requirements.txt
openai>=1.30.0
PyYAML>=6.0
numpy>=1.24
tqdm>=4.66
scikit-learn>=1.3
rich>=13.7
